requirements.txt
Django~=4.2.6
djangorestframework~=3.14.0
asgiref       ~=                3.7.2
Django         ~=               4.2.6
djangorestframework     ~=      3.14.0
djangorestframework-simplejwt ~= 5.3.0
pip                    ~=       23.0.1
PyJWT                   ~=      2.8.0
pytz                 ~=         2023.3.post1
setuptools            ~=        58.1.0
sqlparse               ~=       0.4.4
typing_extensions       ~=      4.8.0
